Rename event dispatcher provider to dispatch post read event provider
